Add OutU* for now
Will add RepOut*,In* later. For now this seems to work (tested with port
0x80). NT seems to have deprecated NtSetInformationProcess with
SE_TCB_PRIVILEGE or whatever starting from Windows 7 and you need a
driver for port I/O, which is easy to write, but definitely not easy to
load (requires signing or booting your PC in some special mode).