x86_64: Make Windows and Linux HCRT's interchangeable
Formerly x86_64_backend relocated FS on Linux and FreeBSD
and GS on Windows as FS and GS are the TLS registers used
in Linux/FreeBSD and Windows respectively - this resulted
in a binary incompatibility that made testing on WINE not
possible. This commit changes it, and allows HCRT2.BIN to
be built from either Windows or Linux or FreeBSD and have
the same ABI on x86_64 architectures.