tls: fix abi clashing with reg w/ inline asm
I am not sure about the semantics of -ffixed
but I am sure that glbl variables are needed
to be consistent across function boundaries.
However the compiler notices x28, a register
that is mandated to be saved, is written to,
and saves it with a stp/ldp pair which means
the asm("x28") global variable is discarded.
So we circumvent by using inline asm and not
showing the compiler that were using the reg
This is actually a bug with Clang, since gcc
removes the store/load pair when -ffixed-x28
is being used. I will file a bug @ llvm-bugs