Commit felony
if (bc_enable) was reading from an inaccesible address that was set to 1
no matter what, even with an __attribute__((constructor)) function that
set bc_enable to 0 because again, accessing it read from that pre-set 1.
removing the bc_enable = 1; line made this stop, so it's something in
the optimizer. even changing it to bc_enable = i; and making
InitBoundsChecker() accept `i` still caused that ghost 1 bug. I don't
know what made the compiler think it was safe to assume bc_enable would
always be 1. It isn't and never has been.

I was forced to write this abomination that masks the variable as a
volatile pointer access. Thankfully the compiler didn't optimize this
away.

Stallman can suck my dick.